requirements.txt
langchain 
langchain-groq 
google-search-results 
python-dotenv 
streamlit